Address Jan's feedback: delete HalfHelpers.netcoreapp.cs, move polyfills to Common
- Delete HalfHelpers.netcoreapp.cs and use #if NET casts in CborReader.Simple.cs
- Move all polyfills from local Polyfills.cs to src/libraries/Common/src/Polyfills/
- Merge BitConverter/DateTimeOffset polyfills with existing common files
- Add new common polyfill files: BigInteger, BinaryPrimitives, Decimal, Stack
- Add string.Create<TState> polyfill to existing StringPolyfills.cs
- Fix BigInteger.ToByteArray polyfill: correct trimming for little-endian + isUnsigned
